requirements.txt
Flask
gunicorn
werkzeug
numpy
tensorflow
Pillow
cloudinary
google-generativeai
python-dotenv
Flask-SQLAlchemy
Flask-Login
Flask-Talisman
email-validator
celery
redis
filetype
Flask-Limiter
pytest
pytest-cov